bug fix started

+++ app/src/main/java/com/example/easy_billing/network/BillingSettingsUpdateRequest.kt
@@ -1,6 +1,6 @@
 package com.example.easy_billing.network
 
 data class BillingSettingsUpdateRequest(
-    val default_gst: Double,
+    val default_gst: Float,
     val printer_layout: String
 )
